Refactor web `createCheckbox` function
