+++ nbdime-web/src/merge/widget/common.ts
@@ -2,6 +2,10 @@
 // Distributed under the terms of the Modified BSD License.
 'use strict';
 
+import {
+  Widget
+} from 'phosphor/lib/ui/widget';
+
 
 // Merge classes:
 export const ONEWAY_LOCAL_CLASS = 'jp-Merge-oneway-local';
@@ -15,4 +19,30 @@ const REMOTE_MERGE_CLASS = 'jp-Merge-remote';
 const MERGED_MERGE_CLASS = 'jp-Merge-merged';
 
 export const MERGE_CLASSES = [BASE_MERGE_CLASS, LOCAL_MERGE_CLASS,
-    REMOTE_MERGE_CLASS, MERGED_MERGE_CLASS];
+  REMOTE_MERGE_CLASS, MERGED_MERGE_CLASS];
+
+
+/**
+ * Create a widget containing a checkbox with a label.
+ *
+ * @export
+ * @param {boolean} value - The initial check state (true = checked)
+ * @param {string} text - The text of the label
+ * @returns {{checkbox: HTMLInputElement, widget: Widget }}
+ */
+export
+function createCheckbox(value: boolean, text: string, indeterminate=false): {checkbox: HTMLInputElement, widget: Widget } {
+  let checkbox = document.createElement('input');
+  checkbox.setAttribute('type', 'checkbox');
+  checkbox.checked = value;
+  checkbox.indeterminate = indeterminate;
+  // Create label for checkbox:
+  let widget = new Widget();
+  let label = document.createElement('label');
+  label.innerHTML = text;
+  // Combine checkbox and label:
+  label.insertBefore(checkbox, label.childNodes[0]);
+  // Add checkbox to header:
+  widget.node.appendChild(label);
+  return {checkbox, widget};
+}
